+++ tests/permissions/test_api_views_inheritance.py
@@ -1,5 +1,5 @@
-from esmerald import APIView, SimpleAPIView, get
-from esmerald.permissions import DenyAll
+from esmerald import APIView, get
+from esmerald.permissions import AllowAny, DenyAll
 from esmerald.testclient import create_client
 
 
@@ -11,8 +11,10 @@ class MyAPIView(APIView):
         return "home"
 
 
-class MySimpleAPIView(SimpleAPIView):
-    @get()
+class MySimpleAPIView(MyAPIView):
+    permissions = [AllowAny]
+
+    @get("/new-home")
     async def get(self) -> str:
         return "home simple"
 
@@ -20,4 +22,10 @@ class MySimpleAPIView(SimpleAPIView):
 def test_cannot_access_apiview(test_client_factory):
     with create_client(routes=[MyAPIView]) as client:
         response = client.get("/home")
+        assert response.status_code == 403
+
+
+def test_cannot_access_simple_apiview(test_client_factory):
+    with create_client(routes=[MySimpleAPIView]) as client:
+        response = client.get("/new-home")
         assert response.status_code == 403